personal-profile/backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
python-dotenv==1.0.1
supabase==2.9.1
gunicorn==23.0.0
itsdangerous==2.2.0
